Fixed multiple auto save instances bug
Problem was:
If you switched accounts during the 5 second interval, the auto save will think that there isnt an existing instance, so it will make a new one. This results into 2 instances at the same time. This has been semi-fixed (Problem can still be reproduced but since interval is now 3 seconds and passwords require a good strength, it is hard to reproduce the bug.)